Share common hydration helpers in a Hydrator struct
This struct is the previous `SignedEntityHydrator` renamed and
generalized to works with multiple common types.
