+++ internal/mithril-persistence/src/database/hydrator.rs
@@ -1,4 +1,4 @@
-//! Signed Entity helpers for persistence
+//! Shared hydrator helpers for persistence
 
 use mithril_common::entities::{
     CardanoDbBeacon, Epoch, SignedEntityType, SignedEntityTypeDiscriminants,
@@ -6,12 +6,37 @@ use mithril_common::entities::{
 
 use crate::sqlite::HydrationError;
 
-/// Helper struct to hydrate [SignedEntityType].
-pub struct SignedEntityTypeHydrator {}
+/// Helper struct to hydrate common data.
+pub struct Hydrator;
 
-impl SignedEntityTypeHydrator {
+impl Hydrator {
+    /// Read a signed entity beacon column from the database
+    pub fn read_signed_entity_beacon_column<U: sqlite::RowIndex + Clone>(
+        row: &sqlite::Row,
+        column_index: U,
+    ) -> String {
+        // We need to check first that the cell can be read as a string first
+        // (e.g. when beacon json is '{"network": "dev", "epoch": 1, "immutable_file_number": 2}').
+        // If it fails, we fallback on reading the cell as an integer (e.g. when beacon json is '5').
+        // TODO: Maybe there is a better way of doing this.
+        match row.try_read::<&str, _>(column_index.clone()) {
+            Ok(value) => value.to_string(),
+            Err(_) => (row.read::<i64, _>(column_index)).to_string(),
+        }
+    }
+
+    /// Try to convert an i64 field from the database to a u64
+    pub fn try_to_u64(field: &str, value: i64) -> Result<u64, HydrationError> {
+        u64::try_from(value)
+            .map_err(|e|
+                HydrationError::InvalidData(
+                    format!("Integer field {field} (value={value}) is incompatible with u64 representation. Error = {e}")
+                )
+            )
+    }
+
     /// Create a [SignedEntityType] from data coming from the database
-    pub fn hydrate(
+    pub fn hydrate_signed_entity_type(
         signed_entity_type_id: usize,
         beacon_str: &str,
     ) -> Result<SignedEntityType, HydrationError> {
