+++ mithril-aggregator/src/database/record/interval_without_block_range_root.rs
@@ -5,9 +5,8 @@ use sqlite::Row;
 
 use mithril_common::entities::BlockNumber;
 use mithril_common::StdResult;
+use mithril_persistence::database::Hydrator;
 use mithril_persistence::sqlite::{HydrationError, Projection, SqLiteEntity};
-
-use crate::database::record::hydrator::try_to_u64;
 
 /// Interval of block numbers (`[start, end[`) without block ranges root.
 pub struct IntervalWithoutBlockRangeRootRecord {
@@ -47,11 +46,11 @@ impl SqLiteEntity for IntervalWithoutBlockRangeRootRecord {
     {
         let start = row
             .read::<Option<i64>, _>(0)
-            .map(|v| try_to_u64("interval_start.start", v))
+            .map(|v| Hydrator::try_to_u64("interval_start.start", v))
             .transpose()?;
         let end = row
             .read::<Option<i64>, _>(1)
-            .map(|v| try_to_u64("interval_end.end", v))
+            .map(|v| Hydrator::try_to_u64("interval_end.end", v))
             .transpose()?;
 
         Ok(Self { start, end })
