+++ mithril-aggregator/src/database/record/open_message.rs
@@ -3,10 +3,8 @@ use sqlite::Row;
 use uuid::Uuid;
 
 use mithril_common::entities::{Epoch, ProtocolMessage, SignedEntityType};
-use mithril_persistence::database::SignedEntityTypeHydrator;
+use mithril_persistence::database::Hydrator;
 use mithril_persistence::sqlite::{HydrationError, Projection, SqLiteEntity};
-
-use crate::database::record::hydrator;
 
 /// ## OpenMessage
 ///
@@ -81,14 +79,14 @@ impl SqLiteEntity for OpenMessageRecord {
         let epoch_setting_id = row.read::<i64, _>(1);
         let epoch_val = u64::try_from(epoch_setting_id)
             .map_err(|e| panic!("Integer field open_message.epoch_setting_id (value={epoch_setting_id}) is incompatible with u64 Epoch representation. Error = {e}"))?;
-        let beacon_str = hydrator::read_signed_entity_beacon_column(&row, 2);
+        let beacon_str = Hydrator::read_signed_entity_beacon_column(&row, 2);
         let signed_entity_type_id = usize::try_from(row.read::<i64, _>(3)).map_err(|e| {
             panic!(
                 "Integer field open_message.signed_entity_type_id cannot be turned into usize: {e}"
             )
         })?;
         let signed_entity_type =
-            SignedEntityTypeHydrator::hydrate(signed_entity_type_id, &beacon_str)?;
+            Hydrator::hydrate_signed_entity_type(signed_entity_type_id, &beacon_str)?;
         let is_certified = row.read::<i64, _>(5) != 0;
         let datetime = &row.read::<&str, _>(7);
         let created_at =
