+++ mithril-aggregator/src/database/record/signed_entity.rs
@@ -10,10 +10,8 @@ use mithril_common::messages::{
 };
 use mithril_common::signable_builder::Artifact;
 use mithril_common::StdError;
-use mithril_persistence::database::SignedEntityTypeHydrator;
+use mithril_persistence::database::Hydrator;
 use mithril_persistence::sqlite::{HydrationError, Projection, SqLiteEntity};
-
-use crate::database::record::hydrator;
 
 /// SignedEntity record is the representation of a stored signed_entity.
 #[derive(Debug, PartialEq, Clone)]
@@ -241,13 +239,13 @@ impl SqLiteEntity for SignedEntityRecord {
         let signed_entity_id = row.read::<&str, _>(0).to_string();
         let signed_entity_type_id_int = row.read::<i64, _>(1);
         let certificate_id = row.read::<&str, _>(2).to_string();
-        let beacon_str = hydrator::read_signed_entity_beacon_column(&row, 3);
+        let beacon_str = Hydrator::read_signed_entity_beacon_column(&row, 3);
         let artifact_str = row.read::<&str, _>(4).to_string();
         let created_at = row.read::<&str, _>(5);
 
         let signed_entity_record = Self {
             signed_entity_id,
-            signed_entity_type: SignedEntityTypeHydrator::hydrate(
+            signed_entity_type: Hydrator::hydrate_signed_entity_type(
                 signed_entity_type_id_int.try_into().map_err(|e| {
                     HydrationError::InvalidData(format!(
                         "Could not cast i64 ({signed_entity_type_id_int}) to u64. Error: '{e}'"
