+++ mithril-signer/src/database/record/cardano_transaction.rs
@@ -3,9 +3,8 @@ use sqlite::Row;
 use mithril_common::entities::{
     BlockHash, BlockNumber, CardanoTransaction, ImmutableFileNumber, SlotNumber, TransactionHash,
 };
+use mithril_persistence::database::Hydrator;
 use mithril_persistence::sqlite::{HydrationError, Projection, SqLiteEntity};
-
-use crate::database::record::hydrator::try_to_u64;
 
 /// Cardano Transaction record is the representation of a cardano transaction.
 #[derive(Debug, PartialEq, Clone)]
@@ -56,11 +55,11 @@ impl SqLiteEntity for CardanoTransactionRecord {
         Self: Sized,
     {
         let transaction_hash = row.read::<&str, _>(0);
-        let block_number = try_to_u64("cardano_tx.block_number", row.read::<i64, _>(1))?;
-        let slot_number = try_to_u64("cardano_tx.slot_number", row.read::<i64, _>(2))?;
+        let block_number = Hydrator::try_to_u64("cardano_tx.block_number", row.read::<i64, _>(1))?;
+        let slot_number = Hydrator::try_to_u64("cardano_tx.slot_number", row.read::<i64, _>(2))?;
         let block_hash = row.read::<&str, _>(3);
         let immutable_file_number =
-            try_to_u64("cardano_tx.immutable_file_number", row.read::<i64, _>(4))?;
+            Hydrator::try_to_u64("cardano_tx.immutable_file_number", row.read::<i64, _>(4))?;
 
         Ok(Self {
             transaction_hash: transaction_hash.to_string(),
